#749 도서별 토론방 화면 구현

+++ android/TodokTodok/app/src/main/java/com/team/todoktodok/data/network/response/discussion/page/ActivatedDiscussion.kt
@@ -15,9 +15,9 @@ data class ActivatedDiscussion(
     val createdAt: String,
     val discussionTitle: String,
     val discussionOpinion: String,
-    val viewCount: Int? = null,
-    val likeCount: Int,
-    val commentCount: Int,
+    val viewCount: Long? = null,
+    val likeCount: Long,
+    val commentCount: Long,
     val isLikedByMe: Boolean,
 ) {
     fun toDomain() =
